Move compare to new source file
Created new source file for compare function.
Moved related functions. Changed path variable
in related files.
