+++ diffkemp/utils.py
@@ -4,6 +4,8 @@ import re
 import sys
 
 LLVM_FUNCTION_REGEX = re.compile(r"^define.*@(\w+)\(", flags=re.MULTILINE)
+# Name of YAML output file created by diffkemp compare command.
+CMP_OUTPUT_FILE = "diffkemp-out.yaml"
 
 
 def get_simpll_build_dir():
